moved components to atomic structure
